requirements.txt
numpy
scikit-learn
pandas
pillow
pymorphy3
transformers
torch